# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/jobs/collect": {
            "post": {
                "description": "对某一工作进入收藏夹",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "收藏工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/def.CollectJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/de-collect": {
            "post": {
                "description": "对某一工作取消收藏",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "取消收藏工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/def.CollectJobRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/list-collections": {
            "get": {
                "description": "列出用户收藏的工作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "列出用户收藏的工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/def.ResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/def.ListJobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/list-uploads": {
            "get": {
                "description": "列出用户已上传的工作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "列出用户已上传的工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/def.ResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/def.ListJobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/list-views": {
            "get": {
                "description": "列出最近浏览的工作",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "列出最近浏览的工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/def.ResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/def.ListJobResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/polish": {
            "post": {
                "description": "对多个工作进行高亮标注",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "批量擦亮工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/def.BatchPolishJobsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/api/jobs/pull-off": {
            "post": {
                "description": "将选中的工作进行下架",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "工作管理"
                ],
                "summary": "批量下架工作",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Bearer 用户令牌",
                        "name": "Authorization",
                        "in": "header",
                        "required": true
                    },
                    {
                        "description": "请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/def.BatchPullOffJobsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        },
        "/login": {
            "post": {
                "description": "用户登录使用JWT验证",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户管理"
                ],
                "summary": "用户登录验证",
                "parameters": [
                    {
                        "description": "请求",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/def.UserRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "token",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/def.ResponseForm"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/msDef.Token"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/def.ResponseForm"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "def.BatchPolishJobsRequest": {
            "type": "object",
            "properties": {
                "jobIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "def.BatchPullOffJobsRequest": {
            "type": "object",
            "properties": {
                "jobIds": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "def.CollectJobRequest": {
            "type": "object",
            "properties": {
                "jobId": {
                    "type": "string"
                }
            }
        },
        "def.ListJobResponse": {
            "type": "object",
            "properties": {
                "jobs": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/mdDef.JobDescription"
                    }
                },
                "length": {
                    "type": "integer"
                }
            }
        },
        "def.ResponseForm": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {},
                "msg": {
                    "type": "string"
                }
            }
        },
        "def.UserRequest": {
            "type": "object",
            "properties": {
                "name": {
                    "type": "string"
                },
                "password": {
                    "type": "string"
                }
            }
        },
        "gorm.DeletedAt": {
            "type": "object",
            "properties": {
                "time": {
                    "type": "string"
                },
                "valid": {
                    "description": "Valid is true if Time is not NULL",
                    "type": "boolean"
                }
            }
        },
        "mdDef.JobDescription": {
            "type": "object",
            "properties": {
                "base_position": {
                    "type": "string"
                },
                "collected_times": {
                    "type": "integer"
                },
                "company_name": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "degree": {
                    "type": "string"
                },
                "deletedAt": {
                    "$ref": "#/definitions/gorm.DeletedAt"
                },
                "id": {
                    "type": "integer"
                },
                "industry": {
                    "type": "string"
                },
                "intern_time_in_months": {
                    "type": "integer"
                },
                "is_show": {
                    "type": "boolean"
                },
                "job_id": {
                    "type": "string"
                },
                "job_name": {
                    "type": "string"
                },
                "job_type": {
                    "type": "string"
                },
                "polished_time": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "upload_user_id": {
                    "type": "string"
                },
                "work_day": {
                    "type": "integer"
                }
            }
        },
        "msDef.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        }
    }
}